@@ -1,44 +1,41 @@
 package subscraping
 
 import (
+	"bytes"
 	"context"
 	"crypto/tls"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/url"
 	"time"
 
 	"github.com/corpix/uarand"
+	"github.com/projectdiscovery/ratelimit"
+
 	"github.com/projectdiscovery/gologger"
-	"go.uber.org/ratelimit"
 )
 
 // NewSession creates a new session object for a domain
-func NewSession(domain string, keys *Keys, proxy string, rateLimit, timeout int, localIP net.IP) (*Session, error) {
-	dialer := &net.Dialer{
-		LocalAddr: &net.TCPAddr{
-			IP: localIP,
-		},
-	}
-
+func NewSession(domain string, proxy string, multiRateLimiter *ratelimit.MultiLimiter, timeout int) (*Session, error) {
 	Transport := &http.Transport{
 		MaxIdleConns:        100,
 		MaxIdleConnsPerHost: 100,
 		TLSClientConfig: &tls.Config{
 			InsecureSkipVerify: true,
 		},
-		DialContext: dialer.DialContext,
+		Dial: (&net.Dialer{
+			Timeout: time.Duration(timeout) * time.Second,
+		}).Dial,
 	}
 
 	// Add proxy
 	if proxy != "" {
 		proxyURL, _ := url.Parse(proxy)
 		if proxyURL == nil {
-			// Log warning but continue anyways
-			gologger.Warning().Msgf("Invalid proxy '%s' provided", proxy)
+			// Log warning but continue anyway
+			gologger.Warning().Msgf("Invalid proxy provided: %s", proxy)
 		} else {
 			Transport.Proxy = http.ProxyURL(proxyURL)
 		}
@@ -49,17 +46,10 @@ func NewSession(domain string, keys *Keys, proxy string, rateLimit, timeout int,
 		Timeout:   time.Duration(timeout) * time.Second,
 	}
 
-	session := &Session{
-		Client: client,
-		Keys:   keys,
-	}
+	session := &Session{Client: client}
 
 	// Initiate rate limit instance
-	if rateLimit > 0 {
-		session.RateLimiter = ratelimit.New(rateLimit)
-	} else {
-		session.RateLimiter = ratelimit.NewUnlimited()
-	}
+	session.MultiRateLimiter = multiRateLimiter
 
 	// Create a new extractor object for the current domain
 	extractor, err := NewSubdomainExtractor(domain)
@@ -112,7 +102,11 @@ func (s *Session) HTTPRequest(ctx context.Context, method, requestURL, cookies s
 		req.Header.Set(key, value)
 	}
 
-	s.RateLimiter.Take()
+	sourceName := ctx.Value(CtxSourceArg).(string)
+	mrlErr := s.MultiRateLimiter.Take(sourceName)
+	if mrlErr != nil {
+		return nil, mrlErr
+	}
 
 	return httpRequestWrapper(s.Client, req)
 }
@@ -120,7 +114,7 @@ func (s *Session) HTTPRequest(ctx context.Context, method, requestURL, cookies s
 // DiscardHTTPResponse discards the response content by demand
 func (s *Session) DiscardHTTPResponse(response *http.Response) {
 	if response != nil {
-		_, err := io.Copy(ioutil.Discard, response.Body)
+		_, err := io.Copy(io.Discard, response.Body)
 		if err != nil {
 			gologger.Warning().Msgf("Could not discard response body: %s\n", err)
 			return
@@ -129,15 +123,27 @@ func (s *Session) DiscardHTTPResponse(response *http.Response) {
 	}
 }
 
+// Close the session
+func (s *Session) Close() {
+	s.MultiRateLimiter.Stop()
+	s.Client.CloseIdleConnections()
+}
+
 func httpRequestWrapper(client *http.Client, request *http.Request) (*http.Response, error) {
-	resp, err := client.Do(request)
+	response, err := client.Do(request)
 	if err != nil {
 		return nil, err
 	}
 
-	if resp.StatusCode != http.StatusOK {
+	if response.StatusCode != http.StatusOK {
 		requestURL, _ := url.QueryUnescape(request.URL.String())
-		return resp, fmt.Errorf("unexpected status code %d received from %s", resp.StatusCode, requestURL)
+
+		gologger.Debug().MsgFunc(func() string {
+			buffer := new(bytes.Buffer)
+			_, _ = buffer.ReadFrom(response.Body)
+			return fmt.Sprintf("Response for failed request against %s:\n%s", requestURL, buffer.String())
+		})
+		return response, fmt.Errorf("unexpected status code %d received from %s", response.StatusCode, requestURL)
 	}
-	return resp, nil
+	return response, nil
 }
