@@ -4,8 +4,7 @@ package sitedossier
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
-	"math/rand"
+	"io"
 	"net/http"
 	"regexp"
 	"time"
@@ -19,68 +18,93 @@ const SleepRandIntn = 5
 
 var reNext = regexp.MustCompile(`<a href="([A-Za-z0-9/.]+)"><b>`)
 
-type agent struct {
-	results chan subscraping.Result
-	session *subscraping.Session
+// Source is the passive scraping agent
+type Source struct {
+	timeTaken time.Duration
+	errors    int
+	results   int
 }
 
-func (a *agent) enumerate(ctx context.Context, baseURL string) {
+// Run function returns all subdomains found with the service
+func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
+	results := make(chan subscraping.Result)
+	s.errors = 0
+	s.results = 0
+
+	go func() {
+		defer func(startTime time.Time) {
+			s.timeTaken = time.Since(startTime)
+			close(results)
+		}(time.Now())
+
+		s.enumerate(ctx, session, fmt.Sprintf("http://www.sitedossier.com/parentdomain/%s", domain), results)
+	}()
+
+	return results
+}
+
+func (s *Source) enumerate(ctx context.Context, session *subscraping.Session, baseURL string, results chan subscraping.Result) {
 	select {
 	case <-ctx.Done():
 		return
 	default:
 	}
 
-	resp, err := a.session.SimpleGet(ctx, baseURL)
+	resp, err := session.SimpleGet(ctx, baseURL)
 	isnotfound := resp != nil && resp.StatusCode == http.StatusNotFound
 	if err != nil && !isnotfound {
-		a.results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
-		a.session.DiscardHTTPResponse(resp)
+		results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
+		s.errors++
+		session.DiscardHTTPResponse(resp)
 		return
 	}
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		a.results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
+		results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
+		s.errors++
 		resp.Body.Close()
 		return
 	}
 	resp.Body.Close()
 
 	src := string(body)
-	for _, match := range a.session.Extractor.FindAllString(src, -1) {
-		a.results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Subdomain, Value: match}
+	for _, subdomain := range session.Extractor.Extract(src) {
+		results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Subdomain, Value: subdomain}
+		s.results++
 	}
 
-	match1 := reNext.FindStringSubmatch(src)
-	time.Sleep(time.Duration((3 + rand.Intn(SleepRandIntn))) * time.Second)
-
-	if len(match1) > 0 {
-		a.enumerate(ctx, "http://www.sitedossier.com"+match1[1])
+	match := reNext.FindStringSubmatch(src)
+	if len(match) > 0 {
+		s.enumerate(ctx, session, fmt.Sprintf("http://www.sitedossier.com%s", match[1]), results)
 	}
-}
-
-// Source is the passive scraping agent
-type Source struct{}
-
-// Run function returns all subdomains found with the service
-func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
-	results := make(chan subscraping.Result)
-
-	a := agent{
-		session: session,
-		results: results,
-	}
-
-	go func() {
-		a.enumerate(ctx, fmt.Sprintf("http://www.sitedossier.com/parentdomain/%s", domain))
-		close(a.results)
-	}()
-
-	return a.results
 }
 
 // Name returns the name of the source
 func (s *Source) Name() string {
 	return "sitedossier"
+}
+
+func (s *Source) IsDefault() bool {
+	return false
+}
+
+func (s *Source) HasRecursiveSupport() bool {
+	return false
+}
+
+func (s *Source) NeedsKey() bool {
+	return false
+}
+
+func (s *Source) AddApiKeys(_ []string) {
+	// no key needed
+}
+
+func (s *Source) Statistics() subscraping.Statistics {
+	return subscraping.Statistics{
+		Errors:    s.errors,
+		Results:   s.results,
+		TimeTaken: s.timeTaken,
+	}
 }
