@@ -8,20 +8,33 @@ import (
 	"time"
 
 	"github.com/hako/durafmt"
+
 	"github.com/projectdiscovery/gologger"
+
+	"github.com/projectdiscovery/subfinder/v2/pkg/passive"
 	"github.com/projectdiscovery/subfinder/v2/pkg/resolve"
 	"github.com/projectdiscovery/subfinder/v2/pkg/subscraping"
 )
 
 const maxNumCount = 2
 
-// EnumerateSingleDomain performs subdomain enumeration against a single domain
-func (r *Runner) EnumerateSingleDomain(ctx context.Context, domain string, outputs []io.Writer) error {
-	gologger.Info().Msgf("Enumerating subdomains for %s\n", domain)
+var replacer = strings.NewReplacer(
+	"/", "",
+	"•.", "",
+	"•", "",
+	"*.", "",
+	"http://", "",
+	"https://", "",
+)
 
-	// Get the API keys for sources from the configuration
-	// and also create the active resolving engine for the domain.
-	keys := r.options.YAMLConfig.GetKeys()
+// EnumerateSingleDomain wraps EnumerateSingleDomainWithCtx with an empty context
+func (r *Runner) EnumerateSingleDomain(domain string, writers []io.Writer) (map[string]map[string]struct{}, error) {
+	return r.EnumerateSingleDomainWithCtx(context.Background(), domain, writers)
+}
+
+// EnumerateSingleDomainWithCtx performs subdomain enumeration against a single domain
+func (r *Runner) EnumerateSingleDomainWithCtx(ctx context.Context, domain string, writers []io.Writer) (map[string]map[string]struct{}, error) {
+	gologger.Info().Msgf("Enumerating subdomains for %s\n", domain)
 
 	// Check if the user has asked to remove wildcards explicitly.
 	// If yes, create the resolution pool and get the wildcards for the current domain
@@ -37,7 +50,7 @@ func (r *Runner) EnumerateSingleDomain(ctx context.Context, domain string, outpu
 
 	// Run the passive subdomain enumeration
 	now := time.Now()
-	passiveResults := r.passiveAgent.EnumerateSubdomains(domain, &keys, r.options.Proxy, r.options.RateLimit, r.options.Timeout, time.Duration(r.options.MaxEnumerationTime)*time.Minute, r.options.LocalIP)
+	passiveResults := r.passiveAgent.EnumerateSubdomainsWithCtx(ctx, domain, r.options.Proxy, r.options.RateLimit, r.options.Timeout, time.Duration(r.options.MaxEnumerationTime)*time.Minute, passive.WithCustomRateLimit(r.rateLimit))
 
 	wg := &sync.WaitGroup{}
 	wg.Add(1)
@@ -45,45 +58,50 @@ func (r *Runner) EnumerateSingleDomain(ctx context.Context, domain string, outpu
 	uniqueMap := make(map[string]resolve.HostEntry)
 	// Create a map to track sources for each host
 	sourceMap := make(map[string]map[string]struct{})
+	skippedCounts := make(map[string]int)
 	// Process the results in a separate goroutine
 	go func() {
 		for result := range passiveResults {
 			switch result.Type {
 			case subscraping.Error:
-				gologger.Warning().Msgf("Could not run source %s: %s\n", result.Source, result.Error)
+				gologger.Warning().Msgf("Encountered an error with source %s: %s\n", result.Source, result.Error)
 			case subscraping.Subdomain:
+				subdomain := replacer.Replace(result.Value)
+
 				// Validate the subdomain found and remove wildcards from
-				if !strings.HasSuffix(result.Value, "."+domain) {
-					continue
-				}
-				subdomain := strings.ReplaceAll(strings.ToLower(result.Value), "*.", "")
-
-				if _, ok := uniqueMap[subdomain]; !ok {
-					sourceMap[subdomain] = make(map[string]struct{})
-				}
-
-				// Log the verbose message about the found subdomain per source
-				if _, ok := sourceMap[subdomain][result.Source]; !ok {
-					gologger.Verbose().Label(result.Source).Msg(subdomain)
-				}
-
-				sourceMap[subdomain][result.Source] = struct{}{}
-
-				// Check if the subdomain is a duplicate. If not,
-				// send the subdomain for resolution.
-				if _, ok := uniqueMap[subdomain]; ok {
+				if !strings.HasSuffix(subdomain, "."+domain) {
+					skippedCounts[result.Source]++
 					continue
 				}
 
-				hostEntry := resolve.HostEntry{Host: subdomain, Source: result.Source}
+				if matchSubdomain := r.filterAndMatchSubdomain(subdomain); matchSubdomain {
+					if _, ok := uniqueMap[subdomain]; !ok {
+						sourceMap[subdomain] = make(map[string]struct{})
+					}
 
-				uniqueMap[subdomain] = hostEntry
+					// Log the verbose message about the found subdomain per source
+					if _, ok := sourceMap[subdomain][result.Source]; !ok {
+						gologger.Verbose().Label(result.Source).Msg(subdomain)
+					}
 
-				// If the user asked to remove wildcard then send on the resolve
-				// queue. Otherwise, if mode is not verbose print the results on
-				// the screen as they are discovered.
-				if r.options.RemoveWildcard {
-					resolutionPool.Tasks <- hostEntry
+					sourceMap[subdomain][result.Source] = struct{}{}
+
+					// Check if the subdomain is a duplicate. If not,
+					// send the subdomain for resolution.
+					if _, ok := uniqueMap[subdomain]; ok {
+						skippedCounts[result.Source]++
+						continue
+					}
+
+					hostEntry := resolve.HostEntry{Domain: domain, Host: subdomain, Source: result.Source}
+
+					uniqueMap[subdomain] = hostEntry
+					// If the user asked to remove wildcard then send on the resolve
+					// queue. Otherwise, if mode is not verbose print the results on
+					// the screen as they are discovered.
+					if r.options.RemoveWildcard {
+						resolutionPool.Tasks <- hostEntry
+					}
 				}
 			}
 		}
@@ -112,38 +130,84 @@ func (r *Runner) EnumerateSingleDomain(ctx context.Context, domain string, outpu
 		}
 	}
 	wg.Wait()
-
-	outputter := NewOutputter(r.options.JSON)
-
+	outputWriter := NewOutputWriter(r.options.JSON)
 	// Now output all results in output writers
 	var err error
-	for _, w := range outputs {
+	for _, writer := range writers {
 		if r.options.HostIP {
-			err = outputter.WriteHostIP(foundResults, w)
+			err = outputWriter.WriteHostIP(domain, foundResults, writer)
 		} else {
 			if r.options.RemoveWildcard {
-				err = outputter.WriteHostNoWildcard(foundResults, w)
+				err = outputWriter.WriteHostNoWildcard(domain, foundResults, writer)
 			} else {
 				if r.options.CaptureSources {
-					err = outputter.WriteSourceHost(sourceMap, w)
+					err = outputWriter.WriteSourceHost(domain, sourceMap, writer)
 				} else {
-					err = outputter.WriteHost(uniqueMap, w)
+					err = outputWriter.WriteHost(domain, uniqueMap, writer)
 				}
 			}
 		}
 		if err != nil {
-			gologger.Error().Msgf("Could not verbose results for %s: %s\n", domain, err)
-			return err
+			gologger.Error().Msgf("Could not write results for %s: %s\n", domain, err)
+			return nil, err
 		}
 	}
 
 	// Show found subdomain count in any case.
 	duration := durafmt.Parse(time.Since(now)).LimitFirstN(maxNumCount).String()
+	var numberOfSubDomains int
 	if r.options.RemoveWildcard {
-		gologger.Info().Msgf("Found %d subdomains for %s in %s\n", len(foundResults), domain, duration)
+		numberOfSubDomains = len(foundResults)
 	} else {
-		gologger.Info().Msgf("Found %d subdomains for %s in %s\n", len(uniqueMap), domain, duration)
+		numberOfSubDomains = len(uniqueMap)
 	}
 
-	return nil
+	if r.options.ResultCallback != nil {
+		if r.options.RemoveWildcard {
+			for host, result := range foundResults {
+				r.options.ResultCallback(&resolve.HostEntry{Domain: host, Host: result.Host, Source: result.Source})
+			}
+		} else {
+			for _, v := range uniqueMap {
+				r.options.ResultCallback(&v)
+			}
+		}
+	}
+	gologger.Info().Msgf("Found %d subdomains for %s in %s\n", numberOfSubDomains, domain, duration)
+
+	if r.options.Statistics {
+		gologger.Info().Msgf("Printing source statistics for %s", domain)
+		statistics := r.passiveAgent.GetStatistics()
+		// This is a hack to remove the skipped count from the statistics
+		// as we don't want to show it in the statistics.
+		// TODO: Design a better way to do this.
+		for source, count := range skippedCounts {
+			if stat, ok := statistics[source]; ok {
+				stat.Results -= count
+				statistics[source] = stat
+			}
+		}
+		printStatistics(statistics)
+	}
+
+	return sourceMap, nil
+}
+
+func (r *Runner) filterAndMatchSubdomain(subdomain string) bool {
+	if r.options.filterRegexes != nil {
+		for _, filter := range r.options.filterRegexes {
+			if m := filter.MatchString(subdomain); m {
+				return false
+			}
+		}
+	}
+	if r.options.matchRegexes != nil {
+		for _, match := range r.options.matchRegexes {
+			if m := match.MatchString(subdomain); m {
+				return true
+			}
+		}
+		return false
+	}
+	return true
 }
