@@ -9,31 +9,40 @@ import (
 	"strings"
 
 	jsoniter "github.com/json-iterator/go"
+
 	"github.com/projectdiscovery/subfinder/v2/pkg/resolve"
 )
 
-// OutPutter outputs content to writers.
-type OutPutter struct {
+// OutputWriter outputs content to writers.
+type OutputWriter struct {
 	JSON bool
 }
 
-type jsonResult struct {
+type jsonSourceResult struct {
 	Host   string `json:"host"`
-	IP     string `json:"ip"`
+	Input  string `json:"input"`
 	Source string `json:"source"`
 }
 
-type jsonSourceResult struct {
+type jsonSourceIPResult struct {
+	Host   string `json:"host"`
+	IP     string `json:"ip"`
+	Input  string `json:"input"`
+	Source string `json:"source"`
+}
+
+type jsonSourcesResult struct {
 	Host    string   `json:"host"`
+	Input   string   `json:"input"`
 	Sources []string `json:"sources"`
 }
 
-// NewOutputter creates a new Outputter
-func NewOutputter(json bool) *OutPutter {
-	return &OutPutter{JSON: json}
+// NewOutputWriter creates a new OutputWriter
+func NewOutputWriter(json bool) *OutputWriter {
+	return &OutputWriter{JSON: json}
 }
 
-func (o *OutPutter) createFile(filename string, appendtoFile bool) (*os.File, error) {
+func (o *OutputWriter) createFile(filename string, appendToFile bool) (*os.File, error) {
 	if filename == "" {
 		return nil, errors.New("empty filename")
 	}
@@ -51,7 +60,7 @@ func (o *OutPutter) createFile(filename string, appendtoFile bool) (*os.File, er
 
 	var file *os.File
 	var err error
-	if appendtoFile {
+	if appendToFile {
 		file, err = os.OpenFile(filename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	} else {
 		file, err = os.Create(filename)
@@ -64,17 +73,17 @@ func (o *OutPutter) createFile(filename string, appendtoFile bool) (*os.File, er
 }
 
 // WriteHostIP writes the output list of subdomain to an io.Writer
-func (o *OutPutter) WriteHostIP(results map[string]resolve.Result, writer io.Writer) error {
+func (o *OutputWriter) WriteHostIP(input string, results map[string]resolve.Result, writer io.Writer) error {
 	var err error
 	if o.JSON {
-		err = writeJSONHostIP(results, writer)
+		err = writeJSONHostIP(input, results, writer)
 	} else {
-		err = writePlainHostIP(results, writer)
+		err = writePlainHostIP(input, results, writer)
 	}
 	return err
 }
 
-func writePlainHostIP(results map[string]resolve.Result, writer io.Writer) error {
+func writePlainHostIP(_ string, results map[string]resolve.Result, writer io.Writer) error {
 	bufwriter := bufio.NewWriter(writer)
 	sb := &strings.Builder{}
 
@@ -96,14 +105,15 @@ func writePlainHostIP(results map[string]resolve.Result, writer io.Writer) error
 	return bufwriter.Flush()
 }
 
-func writeJSONHostIP(results map[string]resolve.Result, writer io.Writer) error {
+func writeJSONHostIP(input string, results map[string]resolve.Result, writer io.Writer) error {
 	encoder := jsoniter.NewEncoder(writer)
 
-	var data jsonResult
+	var data jsonSourceIPResult
 
 	for _, result := range results {
 		data.Host = result.Host
 		data.IP = result.IP
+		data.Input = input
 		data.Source = result.Source
 
 		err := encoder.Encode(&data)
@@ -115,27 +125,27 @@ func writeJSONHostIP(results map[string]resolve.Result, writer io.Writer) error 
 }
 
 // WriteHostNoWildcard writes the output list of subdomain with nW flag to an io.Writer
-func (o *OutPutter) WriteHostNoWildcard(results map[string]resolve.Result, writer io.Writer) error {
+func (o *OutputWriter) WriteHostNoWildcard(input string, results map[string]resolve.Result, writer io.Writer) error {
 	hosts := make(map[string]resolve.HostEntry)
 	for host, result := range results {
-		hosts[host] = resolve.HostEntry{Host: result.Host, Source: result.Source}
+		hosts[host] = resolve.HostEntry{Domain: host, Host: result.Host, Source: result.Source}
 	}
 
-	return o.WriteHost(hosts, writer)
+	return o.WriteHost(input, hosts, writer)
 }
 
 // WriteHost writes the output list of subdomain to an io.Writer
-func (o *OutPutter) WriteHost(results map[string]resolve.HostEntry, writer io.Writer) error {
+func (o *OutputWriter) WriteHost(input string, results map[string]resolve.HostEntry, writer io.Writer) error {
 	var err error
 	if o.JSON {
-		err = writeJSONHost(results, writer)
+		err = writeJSONHost(input, results, writer)
 	} else {
-		err = writePlainHost(results, writer)
+		err = writePlainHost(input, results, writer)
 	}
 	return err
 }
 
-func writePlainHost(results map[string]resolve.HostEntry, writer io.Writer) error {
+func writePlainHost(_ string, results map[string]resolve.HostEntry, writer io.Writer) error {
 	bufwriter := bufio.NewWriter(writer)
 	sb := &strings.Builder{}
 
@@ -153,11 +163,15 @@ func writePlainHost(results map[string]resolve.HostEntry, writer io.Writer) erro
 	return bufwriter.Flush()
 }
 
-func writeJSONHost(results map[string]resolve.HostEntry, writer io.Writer) error {
+func writeJSONHost(input string, results map[string]resolve.HostEntry, writer io.Writer) error {
 	encoder := jsoniter.NewEncoder(writer)
 
+	var data jsonSourceResult
 	for _, result := range results {
-		err := encoder.Encode(result)
+		data.Host = result.Host
+		data.Input = input
+		data.Source = result.Source
+		err := encoder.Encode(data)
 		if err != nil {
 			return err
 		}
@@ -166,23 +180,24 @@ func writeJSONHost(results map[string]resolve.HostEntry, writer io.Writer) error
 }
 
 // WriteSourceHost writes the output list of subdomain to an io.Writer
-func (o *OutPutter) WriteSourceHost(sourceMap map[string]map[string]struct{}, writer io.Writer) error {
+func (o *OutputWriter) WriteSourceHost(input string, sourceMap map[string]map[string]struct{}, writer io.Writer) error {
 	var err error
 	if o.JSON {
-		err = writeSourceJSONHost(sourceMap, writer)
+		err = writeSourceJSONHost(input, sourceMap, writer)
 	} else {
-		err = writeSourcePlainHost(sourceMap, writer)
+		err = writeSourcePlainHost(input, sourceMap, writer)
 	}
 	return err
 }
 
-func writeSourceJSONHost(sourceMap map[string]map[string]struct{}, writer io.Writer) error {
+func writeSourceJSONHost(input string, sourceMap map[string]map[string]struct{}, writer io.Writer) error {
 	encoder := jsoniter.NewEncoder(writer)
 
-	var data jsonSourceResult
+	var data jsonSourcesResult
 
 	for host, sources := range sourceMap {
 		data.Host = host
+		data.Input = input
 		keys := make([]string, 0, len(sources))
 		for source := range sources {
 			keys = append(keys, source)
@@ -197,7 +212,7 @@ func writeSourceJSONHost(sourceMap map[string]map[string]struct{}, writer io.Wri
 	return nil
 }
 
-func writeSourcePlainHost(sourceMap map[string]map[string]struct{}, writer io.Writer) error {
+func writeSourcePlainHost(_ string, sourceMap map[string]map[string]struct{}, writer io.Writer) error {
 	bufwriter := bufio.NewWriter(writer)
 	sb := &strings.Builder{}
 
