@@ -6,13 +6,21 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"strconv"
 	"strings"
+	"time"
 
 	jsoniter "github.com/json-iterator/go"
+
 	"github.com/projectdiscovery/subfinder/v2/pkg/subscraping"
 )
 
-const indexURL = "https://index.commoncrawl.org/collinfo.json"
+const (
+	indexURL     = "https://index.commoncrawl.org/collinfo.json"
+	maxYearsBack = 5
+)
+
+var year = time.Now().Year()
 
 type indexResponse struct {
 	ID     string `json:"id"`
@@ -20,20 +28,28 @@ type indexResponse struct {
 }
 
 // Source is the passive scraping agent
-type Source struct{}
-
-var years = [...]string{"2020", "2019", "2018", "2017"}
+type Source struct {
+	timeTaken time.Duration
+	errors    int
+	results   int
+}
 
 // Run function returns all subdomains found with the service
 func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
 	results := make(chan subscraping.Result)
+	s.errors = 0
+	s.results = 0
 
 	go func() {
-		defer close(results)
+		defer func(startTime time.Time) {
+			s.timeTaken = time.Since(startTime)
+			close(results)
+		}(time.Now())
 
 		resp, err := session.SimpleGet(ctx, indexURL)
 		if err != nil {
 			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+			s.errors++
 			session.DiscardHTTPResponse(resp)
 			return
 		}
@@ -42,10 +58,16 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 		err = jsoniter.NewDecoder(resp.Body).Decode(&indexes)
 		if err != nil {
 			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+			s.errors++
 			resp.Body.Close()
 			return
 		}
 		resp.Body.Close()
+
+		years := make([]string, 0)
+		for i := 0; i < maxYearsBack; i++ {
+			years = append(years, strconv.Itoa(year-i))
+		}
 
 		searchIndexes := make(map[string]string)
 		for _, year := range years {
@@ -75,6 +97,30 @@ func (s *Source) Name() string {
 	return "commoncrawl"
 }
 
+func (s *Source) IsDefault() bool {
+	return false
+}
+
+func (s *Source) HasRecursiveSupport() bool {
+	return false
+}
+
+func (s *Source) NeedsKey() bool {
+	return false
+}
+
+func (s *Source) AddApiKeys(_ []string) {
+	// no key needed
+}
+
+func (s *Source) Statistics() subscraping.Statistics {
+	return subscraping.Statistics{
+		Errors:    s.errors,
+		Results:   s.results,
+		TimeTaken: s.timeTaken,
+	}
+}
+
 func (s *Source) getSubdomains(ctx context.Context, searchURL, domain string, session *subscraping.Session, results chan subscraping.Result) bool {
 	for {
 		select {
@@ -85,6 +131,7 @@ func (s *Source) getSubdomains(ctx context.Context, searchURL, domain string, se
 			resp, err := session.Get(ctx, fmt.Sprintf("%s?url=*.%s", searchURL, domain), "", headers)
 			if err != nil {
 				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
 				session.DiscardHTTPResponse(resp)
 				return false
 			}
@@ -96,14 +143,16 @@ func (s *Source) getSubdomains(ctx context.Context, searchURL, domain string, se
 					continue
 				}
 				line, _ = url.QueryUnescape(line)
-				subdomain := session.Extractor.FindString(line)
-				if subdomain != "" {
-					// fix for triple encoded URL
-					subdomain = strings.ToLower(subdomain)
-					subdomain = strings.TrimPrefix(subdomain, "25")
-					subdomain = strings.TrimPrefix(subdomain, "2f")
+				for _, subdomain := range session.Extractor.Extract(line) {
+					if subdomain != "" {
+						// fix for triple encoded URL
+						subdomain = strings.ToLower(subdomain)
+						subdomain = strings.TrimPrefix(subdomain, "25")
+						subdomain = strings.TrimPrefix(subdomain, "2f")
 
-					results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}
+						results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}
+						s.results++
+					}
 				}
 			}
 			resp.Body.Close()
