@@ -5,12 +5,17 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"strconv"
+	"strings"
+	"time"
 
 	jsoniter "github.com/json-iterator/go"
 
 	// postgres driver
 	_ "github.com/lib/pq"
+
 	"github.com/projectdiscovery/subfinder/v2/pkg/subscraping"
+	contextutil "github.com/projectdiscovery/utils/context"
 )
 
 type subdomain struct {
@@ -19,15 +24,25 @@ type subdomain struct {
 }
 
 // Source is the passive scraping agent
-type Source struct{}
+type Source struct {
+	timeTaken time.Duration
+	errors    int
+	results   int
+}
 
 // Run function returns all subdomains found with the service
 func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
 	results := make(chan subscraping.Result)
+	s.errors = 0
+	s.results = 0
 
 	go func() {
-		defer close(results)
-		count := s.getSubdomainsFromSQL(domain, results)
+		defer func(startTime time.Time) {
+			s.timeTaken = time.Since(startTime)
+			close(results)
+		}(time.Now())
+
+		count := s.getSubdomainsFromSQL(ctx, domain, session, results)
 		if count > 0 {
 			return
 		}
@@ -37,26 +52,58 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 	return results
 }
 
-func (s *Source) getSubdomainsFromSQL(domain string, results chan subscraping.Result) int {
+func (s *Source) getSubdomainsFromSQL(ctx context.Context, domain string, session *subscraping.Session, results chan subscraping.Result) int {
 	db, err := sql.Open("postgres", "host=crt.sh user=guest dbname=certwatch sslmode=disable binary_parameters=yes")
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		return 0
 	}
 
 	defer db.Close()
 
-	pattern := "%." + domain
-	query := `SELECT DISTINCT ci.NAME_VALUE as domain FROM certificate_identity ci
-					  WHERE reverse(lower(ci.NAME_VALUE)) LIKE reverse(lower($1))
-					  ORDER BY ci.NAME_VALUE`
-	rows, err := db.Query(query, pattern)
+	limitClause := ""
+	if all, ok := ctx.Value(contextutil.ContextArg("All")).(contextutil.ContextArg); ok {
+		if allBool, err := strconv.ParseBool(string(all)); err == nil && !allBool {
+			limitClause = "LIMIT 10000"
+		}
+	}
+
+	query := fmt.Sprintf(`WITH ci AS (
+				SELECT min(sub.CERTIFICATE_ID) ID,
+					min(sub.ISSUER_CA_ID) ISSUER_CA_ID,
+					array_agg(DISTINCT sub.NAME_VALUE) NAME_VALUES,
+					x509_commonName(sub.CERTIFICATE) COMMON_NAME,
+					x509_notBefore(sub.CERTIFICATE) NOT_BEFORE,
+					x509_notAfter(sub.CERTIFICATE) NOT_AFTER,
+					encode(x509_serialNumber(sub.CERTIFICATE), 'hex') SERIAL_NUMBER
+					FROM (SELECT *
+							FROM certificate_and_identities cai
+							WHERE plainto_tsquery('certwatch', $1) @@ identities(cai.CERTIFICATE)
+								AND cai.NAME_VALUE ILIKE ('%%' || $1 || '%%')
+								%s
+						) sub
+					GROUP BY sub.CERTIFICATE
+			)
+			SELECT array_to_string(ci.NAME_VALUES, chr(10)) NAME_VALUE
+				FROM ci
+						LEFT JOIN LATERAL (
+							SELECT min(ctle.ENTRY_TIMESTAMP) ENTRY_TIMESTAMP
+								FROM ct_log_entry ctle
+								WHERE ctle.CERTIFICATE_ID = ci.ID
+						) le ON TRUE,
+					ca
+				WHERE ci.ISSUER_CA_ID = ca.ID
+				ORDER BY le.ENTRY_TIMESTAMP DESC NULLS LAST;`, limitClause)
+	rows, err := db.QueryContext(ctx, query, domain)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		return 0
 	}
 	if err := rows.Err(); err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		return 0
 	}
 
@@ -67,10 +114,19 @@ func (s *Source) getSubdomainsFromSQL(domain string, results chan subscraping.Re
 		err := rows.Scan(&data)
 		if err != nil {
 			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+			s.errors++
 			return count
 		}
+
 		count++
-		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: data}
+		for _, subdomain := range strings.Split(data, "\n") {
+			for _, value := range session.Extractor.Extract(subdomain) {
+				if value != "" {
+					results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: value}
+					s.results++
+				}
+			}
+		}
 	}
 	return count
 }
@@ -79,6 +135,7 @@ func (s *Source) getSubdomainsFromHTTP(ctx context.Context, domain string, sessi
 	resp, err := session.SimpleGet(ctx, fmt.Sprintf("https://crt.sh/?q=%%25.%s&output=json", domain))
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		session.DiscardHTTPResponse(resp)
 		return false
 	}
@@ -87,6 +144,7 @@ func (s *Source) getSubdomainsFromHTTP(ctx context.Context, domain string, sessi
 	err = jsoniter.NewDecoder(resp.Body).Decode(&subdomains)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		resp.Body.Close()
 		return false
 	}
@@ -94,7 +152,14 @@ func (s *Source) getSubdomainsFromHTTP(ctx context.Context, domain string, sessi
 	resp.Body.Close()
 
 	for _, subdomain := range subdomains {
-		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain.NameValue}
+		for _, sub := range strings.Split(subdomain.NameValue, "\n") {
+			for _, value := range session.Extractor.Extract(sub) {
+				if value != "" {
+					results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: value}
+					s.results++
+				}
+			}
+		}
 	}
 
 	return true
@@ -103,4 +168,28 @@ func (s *Source) getSubdomainsFromHTTP(ctx context.Context, domain string, sessi
 // Name returns the name of the source
 func (s *Source) Name() string {
 	return "crtsh"
+}
+
+func (s *Source) IsDefault() bool {
+	return true
+}
+
+func (s *Source) HasRecursiveSupport() bool {
+	return true
+}
+
+func (s *Source) NeedsKey() bool {
+	return false
+}
+
+func (s *Source) AddApiKeys(_ []string) {
+	// no key needed
+}
+
+func (s *Source) Statistics() subscraping.Statistics {
+	return subscraping.Statistics{
+		Errors:    s.errors,
+		Results:   s.results,
+		TimeTaken: s.timeTaken,
+	}
 }
