@@ -11,13 +11,15 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 
 	jsoniter "github.com/json-iterator/go"
 
+	"github.com/tomnomnom/linkheader"
+
 	"github.com/projectdiscovery/gologger"
 	"github.com/projectdiscovery/subfinder/v2/pkg/subscraping"
-	"github.com/tomnomnom/linkheader"
 )
 
 type textMatch struct {
@@ -36,20 +38,33 @@ type response struct {
 }
 
 // Source is the passive scraping agent
-type Source struct{}
+type Source struct {
+	apiKeys   []string
+	timeTaken time.Duration
+	errors    int
+	results   int
+	skipped   bool
+}
 
 // Run function returns all subdomains found with the service
 func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
 	results := make(chan subscraping.Result)
+	s.errors = 0
+	s.results = 0
 
 	go func() {
-		defer close(results)
+		defer func(startTime time.Time) {
+			s.timeTaken = time.Since(startTime)
+			close(results)
+		}(time.Now())
 
-		if len(session.Keys.GitHub) == 0 {
+		if len(s.apiKeys) == 0 {
+			gologger.Debug().Msgf("Cannot use the %s source because there was no key defined for it.", s.Name())
+			s.skipped = true
 			return
 		}
 
-		tokens := NewTokenManager(session.Keys.GitHub)
+		tokens := NewTokenManager(s.apiKeys)
 
 		searchURL := fmt.Sprintf("https://api.github.com/search/code?per_page=100&q=%s&sort=created&order=asc", domain)
 		s.enumerate(ctx, searchURL, domainRegexp(domain), tokens, session, results)
@@ -66,23 +81,16 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	}
 
 	token := tokens.Get()
-
-	if token.RetryAfter > 0 {
-		if len(tokens.pool) == 1 {
-			gologger.Verbose().Label(s.Name()).Msgf("GitHub Search request rate limit exceeded, waiting for %d seconds before retry... \n", token.RetryAfter)
-			time.Sleep(time.Duration(token.RetryAfter) * time.Second)
-		} else {
-			token = tokens.Get()
-		}
+	headers := map[string]string{
+		"Accept": "application/vnd.github.v3.text-match+json", "Authorization": "token " + token.Hash,
 	}
-
-	headers := map[string]string{"Accept": "application/vnd.github.v3.text-match+json", "Authorization": "token " + token.Hash}
 
 	// Initial request to GitHub search
 	resp, err := session.Get(ctx, searchURL, "", headers)
 	isForbidden := resp != nil && resp.StatusCode == http.StatusForbidden
 	if err != nil && !isForbidden {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		session.DiscardHTTPResponse(resp)
 		return
 	}
@@ -103,15 +111,17 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	err = jsoniter.NewDecoder(resp.Body).Decode(&data)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		resp.Body.Close()
 		return
 	}
 
 	resp.Body.Close()
 
-	err = proccesItems(ctx, data.Items, domainRegexp, s.Name(), session, results)
+	err = s.proccesItems(ctx, data.Items, domainRegexp, s.Name(), session, results)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors++
 		return
 	}
 
@@ -123,6 +133,7 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 			nextURL, err := url.QueryUnescape(link.URL)
 			if err != nil {
 				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
 				return
 			}
 			s.enumerate(ctx, nextURL, domainRegexp, tokens, session, results)
@@ -130,39 +141,60 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	}
 }
 
-// proccesItems procceses github response items
-func proccesItems(ctx context.Context, items []item, domainRegexp *regexp.Regexp, name string, session *subscraping.Session, results chan subscraping.Result) error {
-	for _, item := range items {
-		// find subdomains in code
-		resp, err := session.SimpleGet(ctx, rawURL(item.HTMLURL))
-		if err != nil {
-			if resp != nil && resp.StatusCode != http.StatusNotFound {
-				session.DiscardHTTPResponse(resp)
+// proccesItems process github response items
+func (s *Source) proccesItems(ctx context.Context, items []item, domainRegexp *regexp.Regexp, name string, session *subscraping.Session, results chan subscraping.Result) error {
+	var wg sync.WaitGroup
+	errChan := make(chan error, len(items))
+
+	for _, responseItem := range items {
+		wg.Add(1)
+		go func(responseItem item) {
+			defer wg.Done()
+
+			// find subdomains in code
+			resp, err := session.SimpleGet(ctx, rawURL(responseItem.HTMLURL))
+			if err != nil {
+				if resp != nil && resp.StatusCode != http.StatusNotFound {
+					session.DiscardHTTPResponse(resp)
+				}
+				errChan <- err
+				return
 			}
+
+			if resp.StatusCode == http.StatusOK {
+				scanner := bufio.NewScanner(resp.Body)
+				for scanner.Scan() {
+					line := scanner.Text()
+					if line == "" {
+						continue
+					}
+					for _, subdomain := range domainRegexp.FindAllString(normalizeContent(line), -1) {
+						results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}
+						s.results++
+					}
+				}
+				resp.Body.Close()
+			}
+
+			// find subdomains in text matches
+			for _, textMatch := range responseItem.TextMatches {
+				for _, subdomain := range domainRegexp.FindAllString(normalizeContent(textMatch.Fragment), -1) {
+					results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}
+					s.results++
+				}
+			}
+		}(responseItem)
+	}
+
+	wg.Wait()
+	close(errChan)
+
+	for err := range errChan {
+		if err != nil {
 			return err
 		}
-
-		if resp.StatusCode == http.StatusOK {
-			scanner := bufio.NewScanner(resp.Body)
-			for scanner.Scan() {
-				line := scanner.Text()
-				if line == "" {
-					continue
-				}
-				for _, subdomain := range domainRegexp.FindAllString(normalizeContent(line), -1) {
-					results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}
-				}
-			}
-			resp.Body.Close()
-		}
-
-		// find subdomains in text matches
-		for _, textMatch := range item.TextMatches {
-			for _, subdomain := range domainRegexp.FindAllString(normalizeContent(textMatch.Fragment), -1) {
-				results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}
-			}
-		}
 	}
+
 	return nil
 }
 
@@ -189,4 +221,29 @@ func domainRegexp(domain string) *regexp.Regexp {
 // Name returns the name of the source
 func (s *Source) Name() string {
 	return "github"
+}
+
+func (s *Source) IsDefault() bool {
+	return false
+}
+
+func (s *Source) HasRecursiveSupport() bool {
+	return false
+}
+
+func (s *Source) NeedsKey() bool {
+	return true
+}
+
+func (s *Source) AddApiKeys(keys []string) {
+	s.apiKeys = keys
+}
+
+func (s *Source) Statistics() subscraping.Statistics {
+	return subscraping.Statistics{
+		Errors:    s.errors,
+		Results:   s.results,
+		TimeTaken: s.timeTaken,
+		Skipped:   s.skipped,
+	}
 }
