@@ -2,57 +2,113 @@
 package securitytrails
 
 import (
+	"bytes"
 	"context"
 	"fmt"
+	"net/http"
 	"strings"
+	"time"
 
 	jsoniter "github.com/json-iterator/go"
+
 	"github.com/projectdiscovery/subfinder/v2/pkg/subscraping"
+	"github.com/projectdiscovery/utils/ptr"
 )
 
 type response struct {
+	Meta struct {
+		ScrollID string `json:"scroll_id"`
+	} `json:"meta"`
+	Records []struct {
+		Hostname string `json:"hostname"`
+	} `json:"records"`
 	Subdomains []string `json:"subdomains"`
 }
 
 // Source is the passive scraping agent
-type Source struct{}
+type Source struct {
+	apiKeys   []string
+	timeTaken time.Duration
+	errors    int
+	results   int
+	skipped   bool
+}
 
 // Run function returns all subdomains found with the service
 func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
 	results := make(chan subscraping.Result)
+	s.errors = 0
+	s.results = 0
 
 	go func() {
-		defer close(results)
+		defer func(startTime time.Time) {
+			s.timeTaken = time.Since(startTime)
+			close(results)
+		}(time.Now())
 
-		if session.Keys.Securitytrails == "" {
+		randomApiKey := subscraping.PickRandom(s.apiKeys, s.Name())
+		if randomApiKey == "" {
+			s.skipped = true
 			return
 		}
 
-		resp, err := session.Get(ctx, fmt.Sprintf("https://api.securitytrails.com/v1/domain/%s/subdomains", domain), "", map[string]string{"APIKEY": session.Keys.Securitytrails})
-		if err != nil {
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-			session.DiscardHTTPResponse(resp)
-			return
-		}
+		var scrollId string
+		headers := map[string]string{"Content-Type": "application/json", "APIKEY": randomApiKey}
 
-		var securityTrailsResponse response
-		err = jsoniter.NewDecoder(resp.Body).Decode(&securityTrailsResponse)
-		if err != nil {
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-			resp.Body.Close()
-			return
-		}
+		for {
+			var resp *http.Response
+			var err error
 
-		resp.Body.Close()
-
-		for _, subdomain := range securityTrailsResponse.Subdomains {
-			if strings.HasSuffix(subdomain, ".") {
-				subdomain += domain
+			if scrollId == "" {
+				var requestBody = []byte(fmt.Sprintf(`{"query":"apex_domain='%s'"}`, domain))
+				resp, err = session.Post(ctx, "https://api.securitytrails.com/v1/domains/list?include_ips=false&scroll=true", "",
+					headers, bytes.NewReader(requestBody))
 			} else {
-				subdomain = subdomain + "." + domain
+				resp, err = session.Get(ctx, fmt.Sprintf("https://api.securitytrails.com/v1/scroll/%s", scrollId), "", headers)
 			}
 
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}
+			if err != nil && ptr.Safe(resp).StatusCode == 403 {
+				resp, err = session.Get(ctx, fmt.Sprintf("https://api.securitytrails.com/v1/domain/%s/subdomains", domain), "", headers)
+			}
+
+			if err != nil {
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
+				session.DiscardHTTPResponse(resp)
+				return
+			}
+
+			var securityTrailsResponse response
+			err = jsoniter.NewDecoder(resp.Body).Decode(&securityTrailsResponse)
+			if err != nil {
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
+				resp.Body.Close()
+				return
+			}
+
+			resp.Body.Close()
+
+			for _, record := range securityTrailsResponse.Records {
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: record.Hostname}
+				s.results++
+			}
+
+			for _, subdomain := range securityTrailsResponse.Subdomains {
+				if strings.HasSuffix(subdomain, ".") {
+					subdomain += domain
+				} else {
+					subdomain = subdomain + "." + domain
+				}
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}
+				s.results++
+			}
+
+			scrollId = securityTrailsResponse.Meta.ScrollID
+
+			if scrollId == "" {
+				break
+			}
 		}
 	}()
 
@@ -62,4 +118,29 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 // Name returns the name of the source
 func (s *Source) Name() string {
 	return "securitytrails"
+}
+
+func (s *Source) IsDefault() bool {
+	return true
+}
+
+func (s *Source) HasRecursiveSupport() bool {
+	return true
+}
+
+func (s *Source) NeedsKey() bool {
+	return true
+}
+
+func (s *Source) AddApiKeys(keys []string) {
+	s.apiKeys = keys
+}
+
+func (s *Source) Statistics() subscraping.Statistics {
+	return subscraping.Statistics{
+		Errors:    s.errors,
+		Results:   s.results,
+		TimeTaken: s.timeTaken,
+		Skipped:   s.skipped,
+	}
 }
