@@ -3,15 +3,34 @@ package subscraping
 import (
 	"context"
 	"net/http"
-	"regexp"
+	"time"
 
-	"go.uber.org/ratelimit"
+	"github.com/projectdiscovery/ratelimit"
+	mapsutil "github.com/projectdiscovery/utils/maps"
 )
+
+type CtxArg string
+
+const (
+	CtxSourceArg CtxArg = "source"
+)
+
+type CustomRateLimit struct {
+	Custom mapsutil.SyncLockMap[string, uint]
+}
 
 // BasicAuth request's Authorization header
 type BasicAuth struct {
 	Username string
 	Password string
+}
+
+// Statistics contains statistics about the scraping process
+type Statistics struct {
+	TimeTaken time.Duration
+	Errors    int
+	Results   int
+	Skipped   bool
 }
 
 // Source is an interface inherited by each passive source
@@ -20,49 +39,42 @@ type Source interface {
 	// which contains the extractor for subdomains, http client
 	// and other stuff.
 	Run(context.Context, string, *Session) <-chan Result
-	// Name returns the name of the source
+
+	// Name returns the name of the source. It is preferred to use lower case names.
 	Name() string
+
+	// IsDefault returns true if the current source should be
+	// used as part of the default execution.
+	IsDefault() bool
+
+	// HasRecursiveSupport returns true if the current source
+	// accepts subdomains (e.g. subdomain.domain.tld),
+	// not just root domains.
+	HasRecursiveSupport() bool
+
+	// NeedsKey returns true if the source requires an API key
+	NeedsKey() bool
+
+	AddApiKeys([]string)
+
+	// Statistics returns the scrapping statistics for the source
+	Statistics() Statistics
+}
+
+// SubdomainExtractor is an interface that defines the contract for subdomain extraction.
+type SubdomainExtractor interface {
+	Extract(text string) []string
 }
 
 // Session is the option passed to the source, an option is created
-// uniquely for eac source.
+// uniquely for each source.
 type Session struct {
-	// Extractor is the regex for subdomains created for each domain
-	Extractor *regexp.Regexp
-	// Keys is the API keys for the application
-	Keys *Keys
+	//SubdomainExtractor
+	Extractor SubdomainExtractor
 	// Client is the current http client
 	Client *http.Client
 	// Rate limit instance
-	RateLimiter ratelimit.Limiter
-}
-
-// Keys contains the current API Keys we have in store
-type Keys struct {
-	Binaryedge           string   `json:"binaryedge"`
-	CensysToken          string   `json:"censysUsername"`
-	CensysSecret         string   `json:"censysPassword"`
-	Certspotter          string   `json:"certspotter"`
-	Chaos                string   `json:"chaos"`
-	Chinaz               string   `json:"chinaz"`
-	DNSDB                string   `json:"dnsdb"`
-	GitHub               []string `json:"github"`
-	IntelXHost           string   `json:"intelXHost"`
-	IntelXKey            string   `json:"intelXKey"`
-	PassiveTotalUsername string   `json:"passivetotal_username"`
-	PassiveTotalPassword string   `json:"passivetotal_password"`
-	Recon                string   `json:"recon"`
-	Robtex               string   `json:"robtex"`
-	Securitytrails       string   `json:"securitytrails"`
-	Shodan               string   `json:"shodan"`
-	Spyse                string   `json:"spyse"`
-	ThreatBook           string   `json:"threatbook"`
-	URLScan              string   `json:"urlscan"`
-	Virustotal           string   `json:"virustotal"`
-	ZoomEyeUsername      string   `json:"zoomeye_username"`
-	ZoomEyePassword      string   `json:"zoomeye_password"`
-	FofaUsername         string   `json:"fofa_username"`
-	FofaSecret           string   `json:"fofa_secret"`
+	MultiRateLimiter *ratelimit.MultiLimiter
 }
 
 // Result is a result structure returned by a source
